Starting project
I'm following great tutorial about building your blockchain
https://hackernoon.com/learn-blockchains-by-building-one-117428612f46

In the first iteration, I'm just going to follow it step by step. The examples are written in python so in the first iteration, I'm not
going to make it Scala like. I'm just going to make compiler happy and learn blockchain concepts first.

Adding new transaction method that keeps in for about sender and recipient and amount that is transacted.