Add new block functionality
From the article:
```
When our Blockchain is instantiated we’ll need to seed it with a genesis block—a block with no predecessors. We’ll also need to add a “proof” to our genesis block which is the result of mining (or proof of work)
```
I added the newBlock function. I'm not 100% sure why do we reset
currentTransactions after creating a new block. Hope I understand it later in the process.